refactor: format large currency values in charts
This commit introduces a new extension function `toLargeValueCurrency` to format large currency values (e.g., 1000 to 1K, 1000000 to 1M) for display in charts.

The following charts have been updated to use this new formatting:
- BudgetVsOutflowChart
- CashflowLineChart
- NetWorthBarChart
- SpendingTrendLineChart

Additionally, the AccountCard has been updated to use this formatting for the account balance. The width of the Y-axis label column in the charts has been increased to accommodate the potentially longer formatted values.

Other minor changes include:
- Using `BigDecimal.ZERO` and `BigDecimal.ONE` for comparisons in `BudgetCategoryCard`.
- Adjusting calculations involving percentages in `HomeScreen` to use `BigDecimal`.
- A minor import change in `AdjustBalanceSheet`.